Cleanup by making monitors/jobs package
